seedjob/requirements.txt
mysqlclient
mysql-connector-python
